Crate identity_diff

This module implements a Diff trait type. The Diff trait gives data structures an ability to compare themselves to another data structure of the same type over time. The library pairs off with identity_derive which implements a derive macro for the Diff Trait. Types supported include HashMap, Option, String, serde_json::Value, Vec and primitives such as i8/u8 up to usize and isize as well as the unit type (), bool, and char types. Structs and Enums are supported via identity_derive and can be composed of any number of these types.

Modules

error

hashmap

hashset

macros

object

option

string

traits

url

value

vec

Structs

DiffHashMap
A DiffHashMap type which represents a Diffed HashMap. By default this value is transparent to serde.

DiffHashSet

DiffString
The Diff Type for a String type.

DiffVec
The Diff Type for Vec.

Enums

DiffOption
A DiffOption<T> type which represents a Diffed Option<T>. By default this value is untagged for serde. It also converts to and from Option when serialized/deserialized

Error

Traits

Diff
The primary Diff Trait type.

Type Definitions

DiffObject

Result